changed the type of parameter sum for start_on_a_new_row()

+++ src/printtext.c
@@ -928,7 +928,7 @@ do_indent(WINDOW *win, const int indent, int *insert_count)
  * Start on a new row?
  */
 static bool
-start_on_a_new_row(const long int sum)
+start_on_a_new_row(const ptrdiff_t sum)
 {
     return (sum < (COLS - 1) ? false : true);
 }
